requirements.txt
dj-database-url==0.5.0
Django==2.0.4
django-bootstrap3==9.1.0
django-filter==1.1.0
django-tables2==1.21.2
gunicorn==19.7.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
